MMA-Bench/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
tqdm>=4.66.0
scikit-learn>=1.3.0
tabulate>=0.9.0
